Move from player objects to UUID clean up deleted Algorithm.java
