+++ src/main/java/net/koozumaa/mazegenerator/MazeGenerator.java
@@ -3,26 +3,18 @@ package net.koozumaa.mazegenerator;
 import net.koozumaa.mazegenerator.Generators.GenManager;
 import net.koozumaa.mazegenerator.Generators.MainGenerator;
 import net.koozumaa.mazegenerator.Tab.TabCompleter;
-import net.koozumaa.mazegenerator.Utils.KoozuPair;
 import net.koozumaa.mazegenerator.Utils.MazeCalcUtils;
 import net.koozumaa.mazegenerator.Utils.PlayerVar;
-import org.bukkit.Location;
-import org.bukkit.Material;
 import org.bukkit.plugin.java.JavaPlugin;
 
 import java.util.ArrayList;
 
 public final class MazeGenerator extends JavaPlugin {
     public static MazeGenerator instance;
-    ArrayList<KoozuPair<Location, Location>> pointList = new ArrayList<>();
     public static String commandPrefix = "§f§lMoti§a§lMaze §8§l» §7";
-    public ArrayList<Material> wallMaterials = new ArrayList<>();
-    public ArrayList<Material> floorMaterial = new ArrayList<>();
-    public ArrayList<Material> roofMaterial = new ArrayList<>();
     public static String commandFormat = "command: /Maze [pos1, pos2, generate, variables]";
     public ArrayList<PlayerVar> playerVars = new ArrayList<>();
 
-    public int blocksPerSecond = 2000;
     public int blocksPerSecondStock = 2000;
     public MazeCalcUtils utils;
     public TabCompleter tabCompleter;
