+++ src/main/resources/plugin.yml
@@ -1,7 +1,7 @@
 name: MazeGenerator
 version: ${project.version}
 main: net.koozumaa.mazegenerator.MazeGenerator
-api-version: 1.16
+api-version: 1.18
 authors: [ Koozu ]
 load: STARTUP
 commands:
